Lista_de_Puntos_with_r_equal_2
This alternative version of the point list is generated for a circle with a radius of 2. This allows the figure's area to fit within the 50 cm^2 limit of the robot's workspace. Therefore, the figure can be drawn more accurately, avoiding any possible overflow problems in the workspace. With this optimized point list, better task efficiency and higher quality in the final drawing result are guaranteed.