SIMULATION
In this simulation, we observed the range of motion of the mechanism using links of 9.5 cm without exceeding the given workspace area of 50 cm². The file is fully editable, so you can adjust the lengths of the links by modifying the values assigned to the variables L1 to L5. By doing so, you can generate a new workspace area that fits within the boundaries of the circular paths traced by the mechanism. This feature allows you to experiment with different link lengths and workspace areas, without putting any strain on the mechanism's movement. With the ability to optimize the workspace and link lengths, you can achieve a more precise and efficient mechanism operation, resulting in high-quality drawings.